Kill the CustomCelery class
use functions instead of methods.

This is required to be able to use celery pytest fixtures so one can
really test celery tasks (especially when a task spawns sub tasks).

one (get_queue_lenth) of the 3 methods has been added as (monkeypatched)
method on the Celery class for the sake of bw compat, but this should really
be removed as well as soon as possible (seems only used in swh-archiver).