Fix tests for the scheduler and the API
adapt the configuration to latest refactorings (especially the 'scheduling_db'
config entry which is now just 'db').

Also refactor the scheduler tests to simplify the test class a bit: constants
used in the test do not need to be in the setUp of the test class, making this
setUp method useless.

Doing so, ensure to use simple dicts as test constants instead of
copy-the-modify them. This is (unit)test file, the dumbest, the better.