Introduce SWH_LOG_TARGET to allow switching between console and journal
The current state of the celery worker/utility logging setup makes us log twice:
the default value of log_console can't be overridden by the cli flag in celery
workers.

The environment variable can be set to journal when running as a systemd
service; this shuts down the console logger, and allows the metadata added to
log messages to pass through.