Use a btree of (task_type, md5(arguments)) to match task arguments
The former index on hash(arguments->'args') has lost relevance as about half the
tasks (the ones for the loader) have the same value (an empty list) for this
field.

This index is more universal, faster, and also easier to convince the planner of
using.

If we want more specific indexes (e.g. on specific keyword arguments) we'll be
able to add that separately.