scheduler.backend_es: Leave index opened when streaming bulk
Prior to this commit, we had the proper behavior of closing index when done
streaming. Unfortunately, this created too much gc on es nodes down the line.
So for now, we remove that behavior.

Note that this implies we need another cog that makes a pass once in a while on
indices to close.

Also, this has been running on production for 2 weeks now and no more gc issues
arose since then.