pytest_plugin: Explicitly set hostname in broker_url for celery TestApp
Since the release of kombu 5.1.0, a warning is now issued when a hostname
is not set in the broker_url config value of a celery app.

That change makes the test_celery_monitor_ping test fails due to that
new unexpected warning.

So explicitly add localhost hostname in the broker_url value of the
celery TestApp config.