recurrent_visits: Allow to set no origins scheduled backoff in config
The send_visits_for_visit_type function uses a default schedule backoff
of 20 minutes where there is no origins to schedule for a given visit
type.

It exists use cases when we would like that schedule backoff to be
shorter in order to schedule listed origins for loading into the
archive more rapidly, typically in the docker environment.

So allow to set that backoff value through configuration.