celery_backend/config: Fix missing comma in setup_log_handler
Because of that missing comma, an exception was raised (tuple object is
not callable) but it was caught and displayed by the _print_errors decorator
so tests could not detect it.

As a consequence, the logging configuration of celery workers was broken.

Add a test to check if an exception was raised by the setup_log_handler
function to avoid bad surprises when deploying to production or in
docker.