requirements-test: Remove swh.lister dependency
The swh.lister package was required as testing dependency to check
registration of celery tasks for listers through plugins declared
in swh.workers entrypoints.

However, it is easy to create a fake lister in tests data to check
its correct registration through scheduler CLI so that swh.lister
dependency is not really needed.

As a consequence, worker plugins are now discovered in the function
register_task_types from the swh.scheduler.cli.task_type module,
previoulsy it was done when importing the module.