Provide a default max_queue_length value to task_type
Because the register task type routine does not provide the value, it's left unchecked.
Once a new lister starts listing origins, the scheduler keeps on scheduling new origins
in the queue without limits. As the consumption may be slower than the production, that
tends towards too much resources usage in rabbitmq.

This should limit the issue for new deployments.