cli/origin: Make utils function compute only the required data
And let the caller function do the actual display. This also make the utility function
compute only the required data to be displayed (the listing data is optionally
outputted).

This also renames the 'list' bool to 'with_listing'. 'list' is a python keyword (it's ok
to reuse but editors usually color it differently than plain function name and that can
be confusing).