scheduler/origin: Add scheduling origins cli from file/stdin to celery
It lived in the snippet repository for years (2017 and prior) and used regularly since
then. It's been recently reused for the osdn origins scheduling. And as usual,
improvments on it ensued. It's time to make it an official cli.

It's now migrated under the the 'swh scheduler origin' subcommand. It's name is
'send-origins-from-file-to-celery'. The file to read is either a local file or the
standard input. This expects a list of origins (urls) to be pushed directly in the
proper queue according to the <task-type> argument.

For example:
```
export SWH_CONFIG_FILENAME=~/.config/swh/scheduler.yml; \
  head -20 /tmp/20230509-1539-priority.list.github | \
  shuf | \
  swh scheduler -C $SWH_CONFIG_FILENAME origin send-origins-from-file-to-celery \
    --queue-name-prefix large-repository \  # optional
    --debug \
    --dry-run \
    load-git
```

or directly:
```
swh scheduler -C $SWH_CONFIG_FILENAME origin send-origins-from-file-to-celery \
    --queue-name-prefix large-repository \  # optional
    --debug \
    --dry-run \
    --limit 10 \
    load-git
    /tmp/20230509-1539-priority.list.github  # note that here, the full file is sent,
                                             # without shuffling it
```

Origins can be routed to extra queue with the help of the <--queue-name-prefix> flag.
This will use standard (configured in the scheduler) queue name with the dedicated
prefix (<queue-name-prefix>:<standard-queue-name>). This also expects that the
destination queue is being consumed on the infra side (ping sysadm for it if not).

The cli can be parametric to limit the number of messages with the <--limit> flag. It
can also just be tested with <--dry-run> to do nothing but print actions. Some extra
logging can be triggered with the <--debug> flag.

Refs. swh/infra/sysadm-environment#4872