Migration psycopg3
The adapter registration become simpler, as psycopg understand enum
natively.

Most change are about the new parameter substitution.

We no longer need to raise Scheduler error if we have multiple Visit
update for the same URl. It seemed fine so I did not make extra work to
make use it happens.

On test was related to psycopg2 page size, we kept it to test "large"
batch.