pika_listener: Add error details to task run on task-failed event
The task-failed event is typically received when a celery worker
exited prematurely.

Add the error traceback in the task run metadata to give more
details about why the task failed.

This is particularly useful for save code now requests as it
gives more details to users (and staffs) about the save request
failure through the Web UI and Web API.

Related to swh/devel/swh-web#4835.