runner: Ensure task runs are created in database prior sending tasks
This is a slight rework of the changes introduced by !407.

This ensures that the task run entries in the scheduler database are
created prior the pika listener set them to started.

Nevertheless task statuses are updated after the tasks were sent to
celery to keep previous behavior.

This fixes some missing task start dates in scheduler database.